comment out owned_by filter for cases sent to Oscar
We expect cases with 2nd referrals to have an `undefined` owned_by user, because we assume that we do NOT want to overwrite the case owner when 2nd referrals come through. This line looks like it was added to prevent syncing cases _without location_, but AK's proposal is to add that location-check validation elsewhere in the job.